Copy code across from the NetBox module to avoid dependency
In order to work around a bug in the Salt native module for NetBox, I've
ported the module over here to be shipped together with salt-sproxy.
The issue is that users bumped into a non-obvious dependency so they
have to run with ``--sync-module`` in order to have the NetBox Roster
pull the data properly.
One solution would be to just default ``--sync-modules`` to True, but
then any modules included in salt-sproxy would implicitly override what
the user has in their own environment - and I don't think that'd be the
best approach, as it may lead to unexpected results.
So I figured the only good approach would be to get rid of the
dependency of the NetBox Roster on the execution module and copy the
necessary bits across.